refactor: Extract WebSocket URL construction to shared utility
Fixes code duplication between app/composables and app/state:
- Created utils/websocket-url.js with buildWebSocketUrl()
- Refactored ConnectionState.js to use shared utility
- Refactored useConnectionState.js to use shared utility
- Removed ~50 lines of duplicated URL construction logic

The WebSocket URL construction (handling port formats like '443/path'
or plain '64738', protocol selection, standard port omission) was
identically duplicated between Knockout and Vue implementations.

Net result: -26 lines duplicate code from 2 files, +52 reusable utility

Tests: 1477 passing ✅
Build: Successful ✅
